fix: fix amount ORAI on table
