+++ src/components/TxList/TransactionTable/TransactionTable.js
@@ -1,40 +1,22 @@
 // @ts-nocheck
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { memo, useMemo } from "react";
+import classNames from "classnames/bind";
+import PropTypes from "prop-types";
+import { memo, useMemo } from "react";
 import { useSelector } from "react-redux";
 import { NavLink } from "react-router-dom";
-import PropTypes from "prop-types";
-import classNames from "classnames/bind";
-import consts from "src/constants/consts";
-import { _, reduceString, setAgoTime, reduceStringAssets } from "src/lib/scripts";
-import { formatOrai, formatFloat, checkTokenCW20 } from "src/helpers/helper";
-import { tableThemes } from "src/constants/tableThemes";
 import ThemedTable from "src/components/common/ThemedTable";
+import consts from "src/constants/consts";
+import { tableThemes } from "src/constants/tableThemes";
+import { checkTokenCW20, formatFloat, formatOrai } from "src/helpers/helper";
+import { getAmountDenomInfo, getNewRoyalty, getRoyaltyAmount, getTokenId, getTxTypeNew } from "src/helpers/transaction";
 import CheckIcon from "src/icons/CheckIcon";
-import TimesIcon from "src/icons/TimesIcon";
 import RedoIcon from "src/icons/RedoIcon";
+import TimesIcon from "src/icons/TimesIcon";
+import { _, reduceString, reduceStringAssets, setAgoTime } from "src/lib/scripts";
 import styles from "./TransactionTable.module.scss";
 
 const cx = classNames.bind(styles);
-
-const handleRoyaltyPercentage = royalty => {
-	if (royalty === "-") {
-		return royalty;
-	}
-
-	royalty = (royalty / consts.ROYALTY_DECIMAL_POINT_PERCENT).toFixed(9);
-	let zeroCount = 0;
-	for (let i = royalty.length - 1; i >= 0; i--) {
-		if (royalty[i] == "0") zeroCount++;
-		if (royalty[i] === ".") {
-			zeroCount++;
-			break;
-		}
-		if (royalty[i] != "0") break;
-	}
-	royalty = royalty.substring(0, royalty.length - zeroCount) + "%";
-	return royalty;
-};
 
 export const getHeaderRow = (royalty = false) => {
 	const txHashHeaderCell = <div className={cx("header-cell", "align-left")}>TxHash</div>;
@@ -75,175 +57,55 @@ export const getHeaderRow = (royalty = false) => {
 	};
 };
 
-export const getNewRoyalty = (account, rawLog = "[]", result = "") => {
-	let newRoyalty = "-";
-	if (result === "Failure") {
-		return newRoyalty;
-	}
-
-	let rawLogArr = JSON.parse(rawLog);
-	let checkRoyalty = false;
-	let checkAccount = false;
-
-	if (rawLogArr && rawLogArr.length > 0) {
-		for (let event of rawLogArr[0]?.events) {
-			if (event["type"] === "wasm") {
-				for (let att of event["attributes"]) {
-					if (att["key"] === "action" && att["value"] === "update_ai_royalty") {
-						checkRoyalty = true;
-						continue;
-					}
-
-					if (checkRoyalty && att["key"] === "creator" && att["value"] === account) {
-						checkAccount = true;
-						continue;
-					}
-
-					if (checkAccount && att["key"] === "new_royalty") {
-						newRoyalty = att["value"];
-						break;
-					}
-				}
-
-				break;
-			}
+export const checkAmountOrai = (denom, amount, newDenom, noDenomName, priceTokens, status, className) => {
+	const denomCheck = checkTokenCW20(denom);
+	let priceToken = status?.price ? " ($" + formatFloat(status.price * (amount / 1000000), 4) + ")" : "";
+	if (denom?.toLowerCase()?.includes("ibc")) {
+		if (newDenom?.toLowerCase()?.includes(consts.GRAVITY)) {
+			denom = newDenom ? reduceStringAssets(newDenom.toLowerCase()) : noDenomName;
+			return (
+				<div className={cx("amount", { "custom-amount": className })}>
+					<span className={cx("amount-value")}>{formatOrai(amount, denomCheck && Math.pow(10, denomCheck.decimal))}</span>
+					<span className={cx("amount-denom")}>{denom}</span>
+					<div className={cx("amount-usd")}>{priceToken}</div>
+				</div>
+			);
+		}
+	} else {
+		if (denom?.toLowerCase()?.includes(consts.GRAVITY)) {
+			denom = newDenom ? reduceStringAssets(newDenom.toLowerCase()) : noDenomName;
+			return (
+				<div className={cx("amount", { "custom-amount": className })}>
+					<span className={cx("amount-value")}>{formatOrai(amount)}</span>
+					<span className={cx("amount-denom")}>{denom}</span>
+					<div className={cx("amount-usd")}>{priceToken}</div>
+				</div>
+			);
 		}
 	}
+	const priceOraiCheck = denom?.toLowerCase() === consts.DENOM_ORAI;
+	const priceTokenCheck = denom?.toLowerCase() === denomCheck?.address?.toLowerCase();
+	const tokenPrice = priceTokens[denomCheck.coingeckoId] || 0;
+	const fmtAmount = !denomCheck.status ? formatOrai(amount) : formatOrai(+amount, Math.pow(10, denomCheck?.decimal), 6);
 
-	return handleRoyaltyPercentage(newRoyalty);
-};
-
-export const getRoyaltyAmount = (account, rawLog = "[]", result = "") => {
-	let royaltyAmount = "0";
-	if (result === "Failure") {
-		return royaltyAmount;
-	}
-
-	let rawLogArr = JSON.parse(rawLog);
-	let checkRoyaltyAmount = false;
-	for (let index = rawLogArr[0]?.events?.length - 1; index > -1; index--) {
-		const event = rawLogArr[0]?.events[index];
-		if (event["type"] === "wasm") {
-			for (let att of event["attributes"]) {
-				if (att["key"] === "action" && att["value"] === "pay_royalty") {
-					checkRoyaltyAmount = true;
-					continue;
-				}
-
-				if (att["key"] === "action" && att["value"] === "finish_pay_royalty") {
-					break;
-				}
-
-				if (checkRoyaltyAmount && att["key"].startsWith(`royalty_${account}_`)) {
-					let index = att["value"].indexOf("orai");
-					royaltyAmount = index !== -1 ? att["value"].slice(0, index) : "0";
-				}
-			}
-
-			break;
-		}
-	}
-
-	const obj = { royalty: checkRoyaltyAmount, amount: royaltyAmount };
-	return obj;
-};
-
-export const getTokenId = (rawLog = "[]", result = "") => {
-	let tokenId = "";
-	if (result === "Failure") {
-		return tokenId;
-	}
-
-	let rawLogArr = JSON.parse(rawLog);
-	for (let event of rawLogArr[0].events) {
-		if (event["type"] === "wasm") {
-			for (let att of event["attributes"]) {
-				if (att["key"] === "token_id") {
-					tokenId = att["value"];
-					break;
-				}
-			}
-
-			break;
-		}
-	}
-
-	return tokenId;
+	return (
+		<div className={cx({ amount: !className }, { "custom-amount": className })}>
+			<span className={cx("amount-value")}>{fmtAmount}</span>
+			<span className={cx("amount-denom")}>{denomCheck.status ? reduceStringAssets(denomCheck.denom) : reduceStringAssets(denom)}</span>
+			{priceOraiCheck ? (
+				<div className={cx("amount-usd")}>{priceToken}</div>
+			) : priceTokenCheck ? (
+				<div className={cx("amount-usd")}>{"($" + formatOrai(+amount * tokenPrice, Math.pow(10, denomCheck.decimal), 6) + ")"}</div>
+			) : (
+				<></>
+			)}
+		</div>
+	);
 };
 
 const TransactionTable = memo(({ data, rowMotions, account, royalty = false, txHashClick = false }) => {
 	const status = useSelector(state => state.blockchain.status);
 	const priceTokens = useSelector(state => state.blockchain.priceTokens);
-
-	const getTxTypeNew = (type, rawLog = "[]", result = "") => {
-		const typeArr = type.split(".");
-		let typeMsg = typeArr[typeArr.length - 1];
-		if (typeMsg === "MsgExecuteContract" && result === "Success") {
-			let rawLogArr = JSON.parse(rawLog);
-			for (let event of rawLogArr[0].events) {
-				if (event["type"] === "wasm") {
-					for (let att of event["attributes"]) {
-						if (att["key"] === "action") {
-							let attValue = att["value"]
-								.split("_")
-								.map(word => word.charAt(0).toUpperCase() + word.slice(1))
-								.join("");
-							typeMsg += "/" + attValue;
-							break;
-						}
-					}
-
-					break;
-				}
-			}
-		}
-		return typeMsg;
-	};
-
-	const checkAmountOrai = (denom, amount, newDenom, noDenomName) => {
-		const denomCheck = checkTokenCW20(denom);
-		let priceToken = status?.price ? " ($" + formatFloat(status.price * (amount / 1000000), 4) + ")" : "";
-		if (denom?.toLowerCase()?.includes("ibc")) {
-			if (newDenom?.toLowerCase()?.includes(consts.GRAVITY)) {
-				denom = newDenom ? reduceStringAssets(newDenom.toLowerCase()) : noDenomName;
-				return (
-					<div className={cx("amount")}>
-						<span className={cx("amount-value")}>{formatOrai(amount, denomCheck && Math.pow(10, denomCheck.decimal))}</span>
-						<span className={cx("amount-denom")}>{denom}</span>
-						<div className={cx("amount-usd")}>{priceToken}</div>
-					</div>
-				);
-			}
-		} else {
-			if (denom?.toLowerCase()?.includes(consts.GRAVITY)) {
-				denom = newDenom ? reduceStringAssets(newDenom.toLowerCase()) : noDenomName;
-				return (
-					<div className={cx("amount")}>
-						<span className={cx("amount-value")}>{formatOrai(amount)}</span>
-						<span className={cx("amount-denom")}>{denom}</span>
-						<div className={cx("amount-usd")}>{priceToken}</div>
-					</div>
-				);
-			}
-		}
-		const priceOraiCheck = denom?.toLowerCase() === consts.DENOM_ORAI;
-		const priceTokenCheck = denom?.toLowerCase() === denomCheck?.address?.toLowerCase();
-		const tokenPrice = priceTokens[denomCheck.coingeckoId] || 0;
-
-		return (
-			<div className={cx("amount")}>
-				<span className={cx("amount-value")}>{denomCheck.status ? formatOrai(+amount, Math.pow(10, denomCheck?.decimal), 6) : formatOrai(amount)}</span>
-				<span className={cx("amount-denom")}>{denomCheck.status ? reduceStringAssets(denomCheck.denom) : reduceStringAssets(denom)}</span>
-				{priceOraiCheck ? (
-					<div className={cx("amount-usd")}>{priceToken}</div>
-				) : priceTokenCheck ? (
-					<div className={cx("amount-usd")}>{"($" + formatOrai(+amount * tokenPrice, Math.pow(10, denomCheck.decimal), 6) + ")"}</div>
-				) : (
-					<></>
-				)}
-			</div>
-		);
-	};
 
 	const getDataRows = (data, txHashStatus) => {
 		if (!Array.isArray(data)) {
@@ -251,9 +113,6 @@ const TransactionTable = memo(({ data, rowMotions, account, royalty = false, txH
 		}
 
 		return data.map(item => {
-			let newDenom = item?.messages?.[0]?.sent_funds?.[0]?.denom_name;
-			let noDenomName = item?.amount?.[0]?.sent_funds?.[0]?.denom;
-			let amountDenomName = item?.amount?.[0]?.amount?.denom;
 			const txHashDataCell = _.isNil(item?.tx_hash) ? (
 				<div className={cx("align-left")}>-</div>
 			) : txHashStatus ? (
@@ -336,8 +195,6 @@ const TransactionTable = memo(({ data, rowMotions, account, royalty = false, txH
 			}
 
 			let amountDataCell;
-			let amount;
-			let denom;
 			const objRoyaltyAmount = getRoyaltyAmount(account, item?.raw_log, item?.result);
 			if (royalty && objRoyaltyAmount.royalty) {
 				amountDataCell = (
@@ -357,30 +214,7 @@ const TransactionTable = memo(({ data, rowMotions, account, royalty = false, txH
 					</div>
 				);
 			} else {
-				if (account && item?.messages?.find(msg => getTxTypeNew(msg["@type"]) === "MsgMultiSend")) {
-					try {
-						let outputs = item?.messages?.[0]?.outputs.find(e => e.address === account);
-						amount = outputs ? outputs?.coins?.[0]?.amount : item?.messages?.[0]?.inputs?.[0]?.coins?.[0]?.amount;
-						denom = outputs ? outputs?.coins?.[0]?.denom : item?.messages?.[0]?.inputs?.[0]?.coins?.[0]?.denom || consts.DENOM;
-					} catch (err) {
-						if (!_.isNil(item?.amount?.[0]?.denom) && !_.isNil(item?.amount?.[0]?.amount)) {
-							amount = item?.amount?.[0]?.amount;
-							denom = item?.amount?.[0]?.denom;
-						}
-					}
-				} else {
-					if (item?.amount?.length > 1) {
-						amount = 0;
-						denom = consts.MORE;
-					} else if (!_.isNil(item?.amount?.[0]?.denom) && !_.isNil(item?.amount?.[0]?.amount)) {
-						denom = newDenom ? newDenom : amountDenomName;
-						amount = item?.amount?.[0]?.amount;
-					} else if (item?.messages[0]?.amount && item?.messages[0]?.amount?.length > 0) {
-						amount = item?.messages[0]?.amount[0]?.amount;
-						denom = item?.messages[0]?.amount[0]?.denom_name;
-					}
-				}
-
+				const { denom, amount, newDenom, noDenomName } = getAmountDenomInfo(item, account);
 				let denomMore;
 				if (denom === consts.MORE) {
 					denomMore = (
@@ -391,14 +225,14 @@ const TransactionTable = memo(({ data, rowMotions, account, royalty = false, txH
 						</div>
 					);
 				} else {
-					denomMore = checkAmountOrai(denom, amount, newDenom, noDenomName);
+					denomMore = checkAmountOrai(denom, amount, newDenom, noDenomName, priceTokens, status);
 				}
 				amountDataCell =
 					_.isNil(amount) || _.isNil(denom) ? (
 						<div className={cx("amount-data-cell", { "amount-data-cell-with-transfer-status": transferStatus }, "align-right")}>
 							{transferStatus && transferStatus}
 							{amount ? (
-								<>{checkAmountOrai(item?.amount?.[0]?.denom, amount, item?.amount?.[0]?.denom_name, noDenomName)}</>
+								<>{checkAmountOrai(item?.amount?.[0]?.denom, amount, item?.amount?.[0]?.denom_name, noDenomName, priceTokens, status)}</>
 							) : (
 								<div className={cx("amount")}>
 									<span className={cx("amount-value")}>0</span>
